Don't create Type{...} for incomplete types in most cases
Core.Typeof is used implicitly in the system, but for types with
free type parameters, e.g. `Vector{Core.TypeVar(:T)}`, it expects
to subsequently be wrapped in a UnionAll. In particular, it is
not in general valid to perform type queries on the result of
`Core.Typeof` (without wrapping it in a UnionAll), but we were
attempting to do so in a number of places. Of course, this is a
bit of a rare sitatuation, since we're not in a habit of passing
around types with free typevars, but since those are valid objects
I think it is better to be defensive.

This introduces `Base.TypeofValid` (better name suggestions welcome),
which is like `Core.Typeof`, except that types with free typevars
get mapped to `DataType`, rather than `Type{}` with a free typevars.
As a result, it is always valid to perform type system operations
on the result of this call.